fix: avatar image url on side bar
